+++ test-module/src/test_error.rs
@@ -1,5 +1,7 @@
 //! Testing error reporting and handling.
 
+use std::fs;
+
 use emacs::{defun, CallEnv, Env, Result, Value};
 use emacs::ErrorKind::{self, Signal, Throw};
 use emacs::ResultExt;
@@ -58,17 +60,16 @@ fn catch<'e>(expected_tag: Value<'e>, lambda: Value<'e>) -> Result<Value<'e>> {
     }
 }
 
-#[allow(deprecated)]
-unsafe fn apply_inner(lambda: Value<'_>, args: Value<'_>) {
+/// Call `apply` on LAMBDA and ARGS, propagating any signaled error.
+#[defun(mod_in_name = false, name = "error:apply")]
+fn apply<'e>(lambda: Value<'e>, args: Value<'e>) -> Result<Value<'e>> {
     let env = lambda.env;
-    env.call("apply", (lambda, args)).unwrap_or_propagate();
+    env.call("apply", (lambda, args))
 }
 
-/// Call `apply` on LAMBDA and ARGS, using panics instead of Result to propagate errors.
-#[defun(mod_in_name = false, name = "error:apply")]
-fn apply(lambda: Value<'_>, args: Value<'_>) -> Result<()> {
-    unsafe { apply_inner(lambda, args); }
-    Ok(())
+#[defun(mod_in_name = false)]
+fn read_file<'e>(env: &Env, path: String) -> Result<String> {
+    fs::read_to_string(path).or_signal(env, emrs_file_error)
 }
 
 #[defun(mod_in_name = false, name = "error:panic")]
@@ -88,8 +89,9 @@ fn parse_arg(env: &CallEnv) -> Result<String> {
 }
 
 emacs::define_errors! {
+    emrs_file_error "File error"
     emacs_module_rs_test_error "Hello" (rust_error)
-    error_defined_without_parent "No error message"
+    error_defined_without_parent "Error"
 }
 
 pub fn init(env: &Env) -> Result<()> {