+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   + Added `use_symbols!`, which enables module code to use Lisp symbols without repeatedly interning them.
   + Added `define_errors!` and `Env::signal` to simplify the process of defining and signaling custom Lisp errors.
 - Raised the minimum supported Rust version to 1.45.
+- Added `ResultExt::or_signal` to make it more convenient to convert a Rust error into a Lisp error.
+- Remove `ResultExt::unwrap_or_propagate`.
 
 ## [0.16.2] - 2021-03-04
 - Fixed compilation on `aarch64-apple-darwin` (Apple Silicon).